moved files to CLI folder, finished commenting
