+++ pipelines/CLI/config/config.ini
@@ -1,0 +1,109 @@
+[default]
+
+pipeline_path       : /approx-vision/pipelines/CLI/core/Pipeline.cpp
+cam_model_path      : /approx-vision/cam_models/NikonD7000/
+wb_index            : 6
+num_ctrl_pts        : 3702
+
+
+###############################################################################
+
+[version]
+
+# Stages Available:
+#     Scale,  
+#     Requant1,       // linear requantize, 
+#     Requant2,       // relies on the fact enum val is 1, 2, 3, ... here
+#     Requant3,
+#     Requant4,
+#     Requant5,
+#     Requant6,
+#     Requant7,
+#     RevToneMap,
+#     RevGamutMap,
+#     RevTransform,
+#     Renoise,
+#     Remosaic,
+#     GaussianBlurCV,
+#     LloydRequant,
+#     Descale,
+#     GamutMap,
+#     Transform,
+#     ToneMap,
+#     DemosSubSample, // subsample demosaic
+#     DemosNN,        // nearest neighbor demosaic
+#     DemosInterp,    // bilinear interpolated demosaic
+#     QrtrResBinning, // quarter resolution pixel binning
+#     PwlToneMap      // piecewise linear tone map
+
+V1  : Scale RevToneMap RevGamutMap RevTransform Renoise Remosaic Descale
+V2  : Scale RevToneMap RevGamutMap RevTransform Renoise Remosaic ToneMap Descale
+V3  : Scale RevToneMap RevGamutMap RevTransform Renoise Remosaic GamutMap Descale
+V4  : Scale RevToneMap RevGamutMap RevTransform Renoise Remosaic Transform Descale
+V5  : Scale RevToneMap RevGamutMap RevTransform Remosaic Descale
+V6  : Scale RevToneMap RevGamutMap RevTransform Renoise Descale
+V7  : Scale RevToneMap Descale
+V8  : Scale RevToneMap RevGamutMap ToneMap Descale
+V9  : Scale RevToneMap RevGamutMap RevTransform GamutMap ToneMap Descale
+V10 : Scale RevToneMap RevGamutMap RevTransform Renoise Transform GamutMap ToneMap Descale
+V11 : Scale RevToneMap RevGamutMap RevTransform Remosaic Transform GamutMap ToneMap Descale
+V12 : Scale RevToneMap RevGamutMap RevTransform ToneMap Descale
+V13 : Scale RevToneMap RevGamutMap RevTransform Renoise ToneMap Descale
+V14 : Scale RevToneMap RevGamutMap RevTransform Renoise Remosaic Descale Requant7
+V15 : Scale RevToneMap RevGamutMap RevTransform Renoise Remosaic Descale Requant6
+V16 : Scale RevToneMap RevGamutMap RevTransform Renoise Remosaic Descale Requant5
+V17 : Scale RevToneMap RevGamutMap RevTransform Renoise Remosaic Descale Requant4
+V18 : Scale RevToneMap RevGamutMap RevTransform Renoise Remosaic Descale Requant3
+V19 : Scale RevToneMap RevGamutMap RevTransform Renoise Remosaic Descale Requant2
+V20 : Scale RevToneMap RevGamutMap RevTransform Renoise Remosaic Descale Requant1
+V21 : Scale RevToneMap RevGamutMap RevTransform Renoise Remosaic ToneMap Descale Requant7
+V22 : Scale RevToneMap RevGamutMap RevTransform Renoise Remosaic ToneMap Descale Requant6
+V23 : Scale RevToneMap RevGamutMap RevTransform Renoise Remosaic ToneMap Descale Requant5
+V24 : Scale RevToneMap RevGamutMap RevTransform Renoise Remosaic ToneMap Descale Requant4
+V25 : Scale RevToneMap RevGamutMap RevTransform Renoise Remosaic ToneMap Descale Requant3
+V26 : Scale RevToneMap RevGamutMap RevTransform Renoise Remosaic ToneMap Descale Requant2
+V27 : Scale RevToneMap RevGamutMap RevTransform Renoise Remosaic ToneMap Descale Requant1
+V28 : Scale RevToneMap RevGamutMap RevTransform Renoise Remosaic LloydRequant Descale
+V29 : Scale RevToneMap RevGamutMap RevTransform Renoise Remosaic LloydRequant Descale Requant7
+V30 : Scale RevToneMap RevGamutMap RevTransform Renoise Remosaic LloydRequant Descale Requant6
+V31 : Scale RevToneMap RevGamutMap RevTransform Renoise Remosaic LloydRequant Descale Requant5
+V32 : Scale RevToneMap RevGamutMap RevTransform Renoise Remosaic LloydRequant Descale Requant4
+V33 : Scale RevToneMap RevGamutMap RevTransform Renoise Remosaic LloydRequant Descale Requant3
+V34 : Scale RevToneMap RevGamutMap RevTransform Renoise Remosaic LloydRequant Descale Requant2
+V35 : Scale RevToneMap RevGamutMap RevTransform Renoise Remosaic LloydRequant Descale Requant1
+V36 : Scale RevToneMap RevGamutMap RevTransform Renoise PwlToneMap Descale
+V37 : Scale RevToneMap RevGamutMap RevTransform Renoise Remosaic PwlToneMap Descale
+V38 : Scale RevToneMap RevGamutMap RevTransform Renoise Remosaic DemosSubSample Descale
+V39 : Scale RevToneMap RevGamutMap RevTransform Renoise Remosaic DemosNN Descale
+V40 : Scale RevToneMap RevGamutMap RevTransform Renoise Remosaic DemosInterp Descale
+V41 : Scale RevToneMap RevGamutMap RevTransform Renoise Remosaic DemosSubSample ToneMap Descale
+V42 : Scale RevToneMap RevGamutMap RevTransform Renoise Remosaic DemosNN ToneMap Descale
+V43 : Scale RevToneMap RevGamutMap RevTransform Renoise Remosaic DemosInterp ToneMap Descale
+V44 : Scale RevToneMap Descale Requant7
+V45 : Scale RevToneMap Descale Requant6
+V46 : Scale RevToneMap Descale Requant5
+V47 : Scale RevToneMap Descale Requant4
+V48 : Scale RevToneMap Descale Requant3
+V49 : Scale RevToneMap Descale Requant2
+V50 : Scale RevToneMap Descale Requant1
+V51 : Requant7
+V52 : Requant6
+V53 : Requant5
+V54 : Requant4
+V55 : Requant3
+V56 : Requant2
+V57 : Requant1
+V58 : Scale RevToneMap RevGamutMap RevTransform Renoise Descale Requant7
+V59 : Scale RevToneMap RevGamutMap RevTransform Renoise Descale Requant6
+V60 : Scale RevToneMap RevGamutMap RevTransform Renoise Descale Requant5
+V61 : Scale RevToneMap RevGamutMap RevTransform Renoise Descale Requant4
+V62 : Scale RevToneMap RevGamutMap RevTransform Renoise Descale Requant3
+V63 : Scale RevToneMap RevGamutMap RevTransform Renoise Descale Requant2
+V64 : Scale RevToneMap RevGamutMap RevTransform Renoise Descale Requant1
+V65 : Scale RevToneMap RevGamutMap RevTransform Renoise ToneMap Descale Requant7
+V66 : Scale RevToneMap RevGamutMap RevTransform Renoise ToneMap Descale Requant6
+V67 : Scale RevToneMap RevGamutMap RevTransform Renoise ToneMap Descale Requant5
+V68 : Scale RevToneMap RevGamutMap RevTransform Renoise ToneMap Descale Requant4
+V69 : Scale RevToneMap RevGamutMap RevTransform Renoise ToneMap Descale Requant3
+V70 : Scale RevToneMap RevGamutMap RevTransform Renoise ToneMap Descale Requant2
+V71 : Scale RevToneMap RevGamutMap RevTransform Renoise ToneMap Descale Requant1
