+++ pipelines/CLI/core/CameraModel.h
