+++ pipelines/CLI/core/PipelineUtil.cpp
@@ -1,5 +1,5 @@
 #include "PipelineUtil.h"
-#include "/approx-vision/pipelines/common_test/CameraModel.h"
+#include "/approx-vision/pipelines/CLI/core/CameraModel.h"
 
 // Pipeline Utility 
 
@@ -113,6 +113,7 @@ Func run_image_pipeline_stage(Func *in_func,
     switch( stage ) {
 
       case Scale: {
+        debug_print("Scale");
         out_func                = make_scale( &out_func );
         break;
       }
@@ -126,12 +127,14 @@ Func run_image_pipeline_stage(Func *in_func,
       }
 
       case RevToneMap: {
+        debug_print("RevToneMap");
         out_func                = make_rev_tone_map(&out_func, rev_tone_h);
         out_func.compute_root();
         break;
       } 
 
       case RevGamutMap: {
+        debug_print("RevGamutMap");
         Func rev_gamut_map_ctrl = make_rbf_ctrl_pts(&out_func, 
                                                     num_ctrl_pts, 
                                                     rev_ctrl_pts_h, 
@@ -145,6 +148,7 @@ Func run_image_pipeline_stage(Func *in_func,
       }
 
       case RevTransform: {
+        debug_print("RevTransform");
         out_func                = make_transform(&out_func, TsTw_tran_inv);
         out_func.compute_root();
         break;
@@ -157,21 +161,25 @@ Func run_image_pipeline_stage(Func *in_func,
 
         switch ( stage ) {
           case Renoise: {
+            debug_print("Renoise");
             OpenCV_renoise(&opencv_in_mat);
             break;
           }
 
           case Remosaic: {
+            debug_print("Remosaic");
             OpenCV_remosaic(&opencv_in_mat);
             break;
           }
 
           case GaussianBlurCV: {
+            debug_print("GaussianBlurCV");
             OpenCV_gaussian_blur(&opencv_in_mat);
             break;
           }
 
           case LloydRequant: {
+            debug_print("LloydRequant");
             OpenCV_lloydmax_requant(&opencv_in_mat);
             break;
           } 
@@ -183,11 +191,13 @@ Func run_image_pipeline_stage(Func *in_func,
       }
 
       case Descale: {
+        debug_print("Descale");
         out_func            = make_descale( &out_func );
         break;
       }
 
       case GamutMap: {
+        debug_print("GamutMap");
         Func gamut_map_ctrl = make_rbf_ctrl_pts(&out_func, 
                                                 num_ctrl_pts, 
                                                 ctrl_pts_h, 
@@ -201,12 +211,14 @@ Func run_image_pipeline_stage(Func *in_func,
       }
 
       case Transform: {
+        debug_print("Transform");
         out_func            = make_transform(&out_func, TsTw_tran);
         out_func.compute_root();
         break;
       }
 
       case ToneMap: {
+        debug_print("ToneMap");
         out_func            = make_tone_map(&out_func, tone_h);
         out_func.compute_root();
         break;
@@ -223,16 +235,19 @@ Func run_image_pipeline_stage(Func *in_func,
         debug_print("clamping image...");
         switch (stage) {
           case DemosSubSample: {
+            debug_print("DemosSubSample");
             out_func            = make_demosaic_subsample(&clamped_func);
             break;
           }
 
           case DemosNN: {
+            debug_print("DemosNN");
             out_func            = make_demosaic_nn(&clamped_func);
             break;
           }
 
           case DemosInterp: {
+            debug_print("DemosInterp");
             out_func            = make_demosaic_interp(&clamped_func);
             break;
           }
@@ -242,6 +257,7 @@ Func run_image_pipeline_stage(Func *in_func,
 
       // changes qrtr_bin_factor implicitly
       case QrtrResBinning: {
+        debug_print("QrtrResBinning");
         out_func            = make_qrtr_res_binning( &out_func );
         debug_print(to_string(qrtr_bin_factor[0]));
         qrtr_bin_factor[0]  = qrtr_bin_factor[0] * 2;
@@ -249,6 +265,7 @@ Func run_image_pipeline_stage(Func *in_func,
       }
 
       case PwlToneMap: {
+        debug_print("PwlToneMap");
         out_func            = make_pwl_tone_map( &out_func );
         break;
       }
@@ -257,228 +274,3 @@ Func run_image_pipeline_stage(Func *in_func,
 
   return out_func;
 }
-
-
-// Func run_image_pipeline_rev(Func *in_func, 
-//                             PipelineStageRev rev_stages[], 
-//                             int num_stages,
-//                             Image<float> *rev_tone_h, 
-//                             int num_ctrl_pts, 
-//                             Image<float> *rev_ctrl_pts_h, 
-//                             Image<float> *rev_weights_h, 
-//                             vector<vector<float>> *rev_coefs, 
-//                             vector<vector<float>> *TsTw_tran_inv ) {
-
-//   Func out_func = *in_func;
-
-//   for (int i = 0; i < num_stages; i++) {
-//     PipelineStageRev stage = rev_stages[i];
-//     debug_print("rev stage: " + to_string(stage));
-
-//     switch( stage ) {
-//       case RevScale: {
-//         out_func                = make_scale( &out_func );
-//         break;
-//       }
-
-//       case RevDescale: {
-//         out_func                = make_descale( &out_func );
-//         break;
-//       }
-
-//       case RevRequant1: case RevRequant2: case RevRequant3: case RevRequant4: 
-//       case RevRequant5: case RevRequant6: case RevRequant7: {
-//         int num_bits            = stage;
-//         out_func                = make_requant(&out_func, num_bits);
-//         debug_print("requant num bits: " + to_string(num_bits));
-//         break;
-//       }
-
-//       case RevToneMap: {
-//         out_func                = make_rev_tone_map(&out_func, rev_tone_h);
-//         out_func.compute_root();
-//         break;
-//       } 
-
-//       case RevGamutMap: {
-//         Func rev_gamut_map_ctrl = make_rbf_ctrl_pts(&out_func, 
-//                                                     num_ctrl_pts, 
-//                                                     rev_ctrl_pts_h, 
-//                                                     rev_weights_h );
-
-//         out_func                = make_rbf_biases(  &out_func, 
-//                                                     &rev_gamut_map_ctrl, 
-//                                                     rev_coefs);
-//         rev_gamut_map_ctrl.compute_root();
-//         break;
-//       }
-
-//       case RevTransform: {
-//         out_func                = make_transform(&out_func, TsTw_tran_inv);
-//         out_func.compute_root();
-//         break;
-//       }
-
-//       case DemosSubSample: {
-//         out_func            = make_demosaic_subsample(&out_func);
-//         break;
-//       }
-
-//       case DemosNN: {
-//         out_func            = make_demosaic_nn(&out_func);
-//         break;
-//       }
-
-//       case DemosInterp: {
-//         out_func            = make_demosaic_interp( &out_func );
-//         break;
-//       }
-
-//       case QrtrResBinning: {
-//         out_func            = make_qrtr_res_binning( &out_func );
-//         debug_print(to_string(qrtr_bin_factor[0]));
-//         qrtr_bin_factor[0]  = qrtr_bin_factor[0] * 2;
-//         break;
-//       }
-
-//       case PwlToneMap: {
-//         out_func            = make_pwl_tone_map( &out_func );
-//         break;
-//       }
-
-//     }
-//   }
-
-//   return out_func;
-// }
-
-// void run_image_pipeline_cv( Mat *InMat, PipelineStageCV cv_stages[], int num_stages ) {
-
-//   using namespace Halide;
-//   using namespace Halide::Tools;
-//   using namespace cv;
-
-//   for (int i = 0; i < num_stages; i++) {
-//     PipelineStageCV stage = cv_stages[i];
-//     debug_print("cv stage:  " + to_string(stage));
-
-//     switch( stage ) {
-//       case Renoise: {
-//         OpenCV_renoise(InMat);
-//         break;
-//       }
-
-//       case Remosaic: {
-//         OpenCV_remosaic(InMat);
-//         break;
-//       }
-
-//       case GaussianBlurCV: {
-//         OpenCV_gaussian_blur(InMat);
-//         break;
-//       }
-
-//       case LloydRequant: {
-//         OpenCV_lloydmax_requant(InMat);
-//         break;
-//       }
-//     }
-//   }
-// }
-
-// Func run_image_pipeline_fwd(Func *in_func, 
-//                             PipelineStageFwd fwd_stages[], 
-//                             int num_stages,
-//                             vector<int> &qrtr_bin_factor,
-//                             Image<float> *tone_h, 
-//                             int num_ctrl_pts, 
-//                             Image<float> *ctrl_pts_h, 
-//                             Image<float> *weights_h, 
-//                             vector<vector<float>> *coefs, 
-//                             vector<vector<float>> *TsTw_tran ) {
-
-//   using namespace Halide;
-//   using namespace Halide::Tools;
-//   using namespace cv;
-
-//   Func out_func = *in_func;
-
-//   for (int i = 0; i < num_stages; i++) {
-//     PipelineStageFwd stage = fwd_stages[i];
-//     debug_print("fwd stage: " + to_string(stage));
-
-//     switch( stage ) {
-//       case Scale: {
-//         out_func            = make_scale( &out_func );
-//         break;
-//       }
-
-//       case Descale: {
-//         out_func            = make_descale( &out_func );
-//         break;
-//       }
-
-//       case ToneMap: {
-//         out_func            = make_tone_map(&out_func, tone_h);
-//         out_func.compute_root();
-//         break;
-//       } 
-
-//       case GamutMap: {
-//         Func gamut_map_ctrl = make_rbf_ctrl_pts(&out_func, 
-//                                                 num_ctrl_pts, 
-//                                                 ctrl_pts_h, 
-//                                                 weights_h);
-
-//         out_func            = make_rbf_biases ( &out_func, 
-//                                                 &gamut_map_ctrl, 
-//                                                 coefs);
-//         gamut_map_ctrl.compute_root();
-//         break;
-//       }
-
-//       case Transform: {
-//         out_func            = make_transform(&out_func, TsTw_tran);
-//         out_func.compute_root();
-//         break;
-//       }
-
-//       case Requant1: case Requant2: case Requant3: case Requant4: 
-//       case Requant5: case Requant6: case Requant7: {
-//         int num_bits        = stage;
-//         out_func            = make_requant(&out_func, num_bits);
-//         debug_print(num_bits);
-//         break;
-//       }
-
-//       case DemosSubSample: {
-//         out_func            = make_demosaic_subsample(&out_func);
-//         break;
-//       }
-
-//       case DemosNN: {
-//         out_func            = make_demosaic_nn(&out_func);
-//         break;
-//       }
-
-//       case DemosInterp: {
-//         out_func            = make_demosaic_interp( &out_func );
-//         break;
-//       }
-
-//       case QrtrResBinning: {
-//         out_func            = make_qrtr_res_binning( &out_func );
-//         debug_print(to_string(qrtr_bin_factor[0]));
-//         qrtr_bin_factor[0]  = qrtr_bin_factor[0] * 2;
-//         break;
-//       }
-
-//       case PwlToneMap: {
-//         out_func            = make_pwl_tone_map( &out_func );
-//         break;
-//       }
-//     }
-//   }
-
-//   return out_func;
-// }
