+++ pipelines/CLI/core/PipelineUtil.h
@@ -1,13 +1,14 @@
 
 #include <stdio.h>
 #include <math.h>
+#include <map>
 
 #include "/approx-vision/pipelines/common/pipe_stages.h"
 #include "/approx-vision/pipelines/common/ImgPipeConfig.h"
 #include "/approx-vision/pipelines/common/LoadCamModel.h"
 #include "/approx-vision/pipelines/common/MatrixOps.h"
 
-#include "/approx-vision/pipelines/common_test/CameraModel.h"
+#include "/approx-vision/pipelines/CLI/core/CameraModel.h"
 
 using namespace std;
 
@@ -53,6 +54,7 @@ enum PipelineStage {
   QrtrResBinning, // quarter resolution pixel binning
   PwlToneMap      // piecewise linear tone map
 };
+
 
 /**
   Runs image file through specified reverse, cv, forward pipeline stages
