+++ pipelines/CLI/legacy/CameraModel.h
