+++ pipelines/CLI/legacy/Makefile
