+++ pipelines/CLI/legacy/pipeline_V7.cpp
