+++ pipelines/CLI/legacy/pipeline_V77.cpp
