+++ pipelines/CLI/legacy/pipeline_V78.cpp
