+++ pipelines/CLI/temp/output.png
