+++ pipelines/CLI/temp/output/example.png
